mcp-acme/requirements.txt
fastapi==0.114.2
uvicorn==0.30.6
psycopg2-binary==2.9.9
requests==2.32.3
cryptography==43.0.1
python-dateutil==2.9.0.post0
pydantic-core!=2.41.3
pydantic>=2.7,<3
